+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-05-22 11:32:26
-// Commit: perf(perf): optimize memory consumption in perf list viewport
+// Updated at: 2026-05-22 14:11:36
+// Commit: perf(perf): apply RepaintBoundary on animated perf widgets
 
 class perfComponentState {
-  final String commitRef = '1312';
-  final String lastMessage = 'perf(perf): optimize memory consumption in perf list viewport';
-  final DateTime timestamp = DateTime.parse('2026-05-22T11:32:26+05:00');
+  final String commitRef = '1319';
+  final String lastMessage = 'perf(perf): apply RepaintBoundary on animated perf widgets';
+  final DateTime timestamp = DateTime.parse('2026-05-22T14:11:36+05:00');
 }